requirements.txt
pony
python-telegram-bot
psycopg2
